Robyn Lang's Coding Dojo final project.
This final project is a website I created for my church choir, Voices of Inspiration. Non-logged in users can view a list of upcoming choir performances with locations, dates, and times; view past performances that open in a new tab; view general information about the choir and its director; view a list of current members by role; donate; send contact requests to choir admins; purchase choir merchandise (product backlog, once merchandise is available); and join the choir via the website. Logged in users can view a list of upcoming events and the songs that need to be learned for them (which also link to videos of the songs that open in a new tab); create song suggestions and upload them to the database; view a list of the songs they have suggested on their dashboard; and edit those suggestions. Also, as long as they remain logged in, they can freely navigate the site and still return to their dashboard without having to log in again. Logged in routes are protected. Validation messages flash when incomplete information is attempted to be sent to the database.